+++ source/extensions/stdapi/server/railgun/railgun.c
@@ -585,7 +585,7 @@ DWORD request_railgun_api( Remote * pRemote, Packet * pPacket )
 			packet_add_tlv_string( pResponse, TLV_TYPE_RAILGUN_BACK_MSG, pErrorMsg );
 		}
 
-		dwResult = PACKET_TRANSMIT( pRemote, pResponse, NULL );
+		dwResult = packet_transmit( pRemote, pResponse, NULL );
 	}
 
 	if( rInput.pBufferIN )
@@ -656,7 +656,7 @@ DWORD request_railgun_memread( Remote * pRemote, Packet * pPacket )
 		if( pData )
 			packet_add_tlv_raw( pResponse, TLV_TYPE_RAILGUN_MEM_DATA, pData, dwLength );
 
-		dwResult = PACKET_TRANSMIT( pRemote, pResponse, NULL );
+		dwResult = packet_transmit( pRemote, pResponse, NULL );
 	}
 
 	if( pData )
@@ -713,7 +713,7 @@ DWORD request_railgun_memwrite( Remote * pRemote, Packet * pPacket )
 	{
 		packet_add_tlv_uint( pResponse, TLV_TYPE_RESULT, dwResult );
 
-		dwResult = PACKET_TRANSMIT( pRemote, pResponse, NULL );
+		dwResult = packet_transmit( pRemote, pResponse, NULL );
 	}
 
 	dprintf("[RAILGUN] request_railgun_memwrite: Finished.");
