--- source/server/win/server_transport_tcp.h
@@ -1,6 +1,0 @@
-#ifndef _METERPRETER_SERVER_SETUP_TCP
-#define _METERPRETER_SERVER_SETUP_TCP
-
-Transport* transport_create_tcp(wchar_t* url, TimeoutSettings* timeouts);
-
-#endif
