+++ workspace/metsrv/metsrv.vcxproj
@@ -656,8 +656,7 @@ copy /y "$(TargetDir)$(TargetFileName)" "$(ProjectDir)..\..\output\$(PlatformSho
     <ClCompile Include="..\..\source\server\win\remote_dispatch.c" />
     <ClCompile Include="..\..\source\server\remote_dispatch_common.c" />
     <ClCompile Include="..\..\source\server\server_setup_win.c" />
-    <ClCompile Include="..\..\source\server\win\server_transport_tcp.c" />
-    <ClCompile Include="..\..\source\server\win\server_transport_winhttp.c">
+    <ClCompile Include="..\..\source\server\win\server_setup_winhttp.c">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='r7_debug|Win32'">NotUsing</PrecompiledHeader>
       <PrecompiledHeaderFile Condition="'$(Configuration)|$(Platform)'=='r7_debug|Win32'">
       </PrecompiledHeaderFile>
@@ -707,8 +706,7 @@ copy /y "$(TargetDir)$(TargetFileName)" "$(ProjectDir)..\..\output\$(PlatformSho
     <ClInclude Include="..\..\source\server\libloader.h" />
     <ClInclude Include="..\..\source\server\metsrv.h" />
     <ClInclude Include="..\..\source\server\remote_dispatch.h" />
-    <ClInclude Include="..\..\source\server\win\server_transport_tcp.h" />
-    <ClInclude Include="..\..\source\server\win\server_transport_winhttp.h" />
+    <ClInclude Include="..\..\source\server\win\server_setup_winhttp.h" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\backcompat\backcompat.vcxproj">
